+++ AlchemistCode/Cards/Basic/Prime.cs
@@ -5,6 +5,7 @@ using BaseLib.Utils;
 using MegaCrit.Sts2.Core.Entities.Cards;
 using MegaCrit.Sts2.Core.GameActions.Multiplayer;
 using MegaCrit.Sts2.Core.Models;
+using MegaCrit.Sts2.Core.ValueProps;
 
 namespace Alchemist.AlchemistCode.Cards.Basic;
 
@@ -14,7 +15,9 @@ public class Prime : AlchemistCard, ITranscendenceCard
 
     public Prime() : base(1, CardType.Skill, CardRarity.Basic, TargetType.Self)
     {
-        WithBlock(4, 3);
+        // The multiplier is 0 or 1, so the face shows the Reaction bonus only while it would land
+        WithCalculatedBlock(4, 2, static (card, _) =>
+            ((AlchemistCard)card).ReactionActive ? 1 : 0, ValueProp.Move, 3, 0);
         WithTips(_ => Infusion.InfuseTips());
     }
 
@@ -24,8 +27,7 @@ public class Prime : AlchemistCard, ITranscendenceCard
 
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {
-        var reacted = ReactionActive;
         await CommonActions.CardBlock(this, play);
-        await Infusion.InfuseChosen(choiceContext, this, PileType.Hand, reacted ? 2 : 1);
+        await Infusion.InfuseChosen(choiceContext, this, PileType.Hand, 1);
     }
 }