+++ AlchemistCode/Cards/Common/Percolate.cs
@@ -9,7 +9,7 @@ public class Percolate : AlchemistCard
 {
     public Percolate() : base(1, CardType.Skill, CardRarity.Common, TargetType.Self)
     {
-        WithVar("draw", 3, 1);
+        WithVar("draw", 2, 1);
         WithTip(typeof(DrawCardsNextTurnPower));
     }