+++ AlchemistCode/Cards/Rare/Eureka.cs
@@ -1,9 +1,8 @@
 using BaseLib.Utils;
-using MegaCrit.Sts2.Core.CardSelection;
-using MegaCrit.Sts2.Core.Commands;
 using MegaCrit.Sts2.Core.Entities.Cards;
 using MegaCrit.Sts2.Core.GameActions.Multiplayer;
 using Alchemist.AlchemistCode.Cards.Token;
+using Alchemist.AlchemistCode.Commands;
 
 namespace Alchemist.AlchemistCode.Cards.Rare;
 
@@ -18,14 +17,7 @@ public class Eureka : AlchemistCard
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {
         await CommonActions.Draw(this, choiceContext);
-        var selected = await CardSelectCmd.FromHand(
-            choiceContext, Owner,
-            new CardSelectorPrefs(CardSelectorPrefs.TransformSelectionPrompt, 1),
-            null, this);
-        foreach (var card in selected)
-        {
-            var distillate = CombatState!.CreateCard<Distillate>(Owner);
-            await CardCmd.Transform(card, distillate);
-        }
+        // The shared helper carries the upgrade onto the Distillate, which the description promises
+        await AlchemistCardCmd.TransformFromHand<Distillate>(choiceContext, this);
     }
 }